pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pgm-scrna"
version = "0.1.0"
description = "Heterogeneity-aware soft-weighted GGM with STRING knowledge-graph priors for scRNA-seq"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "PGM Research" }]
dependencies = [
    "anndata>=0.10",
    "scanpy>=1.10",
    "numpy>=1.26",
    "scipy>=1.11",
    "pandas>=2.1",
    "scikit-learn>=1.3",
    "networkx>=3.2",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.18",
    "tqdm>=4.66",
    "joblib>=1.3",
    "pyarrow>=14",
    "statsmodels>=0.14",
    "pydantic>=2.5",
    "PyYAML>=6.0",
    "rich>=13.7",
    "tenacity>=8.2",
    "requests>=2.31",
    "igraph>=0.11",
    "leidenalg>=0.10",
]

[project.optional-dependencies]
eda = ["ydata-profiling>=4.16; python_version<'3.14'", "sweetviz>=2.3; python_version<'3.14'"]
test = ["pytest>=7.4", "pytest-cov>=4.1"]
dev = ["pgm-scrna[eda,test]", "ruff>=0.4", "jupyter>=1.0", "ipykernel>=6.29"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = ["smoke: fast subset regression"]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
